requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
openai==1.31.2
packaging==24.0
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.1
